# internal/api/openapi.yaml
openapi: "3.1.0"
info:
  title: Ella IMS API
  version: "1.0.0"
  description: |
    Success: `{"result": <payload>}`. Error: `{"error": "message"}`.

    Settings are replaced in full by `PUT`, which answers with the resource as stored. Unknown fields are
    rejected. A change applies without a restart of the process: a change of MCC or MNC, of the Diameter peers or
    of the policy function restarts Diameter and SIP, which peers and UEs reconnect to.
servers:
  - url: "{scheme}://{host}"
    variables:
      scheme:
        default: http
        enum: [http, https]
      host:
        default: localhost:5020
security: []
...
paths:
...
  /api/v1/registrations/{impi}/reauthenticate:
    post:
      tags: [Registrations]
      operationId: reauthenticate
      summary: Re-authenticate a registered UE
      parameters:
        - name: impi
          in: path
          required: true
          schema:
            type: string
      responses:
        "202":
          description: Re-authentication started
          content:
            application/json:
              schema:
                type: object
                properties:
                  result:
                    type: object
                    properties:
                      impi:
                        type: string
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
  responses:
...
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    InternalError:
      description: Internal error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"